+++ app/metashell/main.cpp
@@ -21,6 +21,7 @@
 #include <metashell/engine_entry.hpp>
 #include <metashell/engine_gcc.hpp>
 #include <metashell/engine_null.hpp>
+#include <metashell/engine_vc.hpp>
 #include <metashell/fstream_file_writer.hpp>
 #include <metashell/logger.hpp>
 #include <metashell/make_unique.hpp>
@@ -76,7 +77,8 @@ int main(int argc_, const char* argv_[])
         {"internal", metashell::get_internal_templight_entry()},
         {"clang", metashell::get_engine_clang_entry()},
         {"null", metashell::get_engine_null_entry()},
-        {"gcc", metashell::get_engine_gcc_entry()}};
+        {"gcc", metashell::get_engine_gcc_entry()},
+        {"msvc", metashell::get_engine_vc_entry()}};
 
     metashell::default_environment_detector det(argv_[0]);