+++ include/metashell/engine_vc.hpp
@@ -1,5 +1,5 @@
-#ifndef METASHELL_PREPROCESSOR_SHELL_CONSTANT_HPP
-#define METASHELL_PREPROCESSOR_SHELL_CONSTANT_HPP
+#ifndef METASHELL_ENGINE_VC_HPP
+#define METASHELL_ENGINE_VC_HPP
 
 // Metashell - Interactive C++ template metaprogramming shell
 // Copyright (C) 2016, Abel Sinkovics (abel@sinkovics.hu)
@@ -17,22 +17,11 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <metashell/iface/preprocessor_shell.hpp>
-
-#include <string>
+#include <metashell/engine_entry.hpp>
 
 namespace metashell
 {
-  class preprocessor_shell_constant : public iface::preprocessor_shell
-  {
-  public:
-    explicit preprocessor_shell_constant(data::result result_);
-
-    virtual data::result precompile(const std::string&) override;
-
-  private:
-    data::result _result;
-  };
+  engine_entry get_engine_vc_entry();
 }
 
 #endif