+++ include/metashell/macro_discovery_clang.hpp
@@ -1,5 +1,5 @@
-#ifndef METASHELL_PREPROCESSOR_SHELL_CONSTANT_HPP
-#define METASHELL_PREPROCESSOR_SHELL_CONSTANT_HPP
+#ifndef METASHELL_MACRO_DISCOVERY_CLANG_HPP
+#define METASHELL_MACRO_DISCOVERY_CLANG_HPP
 
 // Metashell - Interactive C++ template metaprogramming shell
 // Copyright (C) 2016, Abel Sinkovics (abel@sinkovics.hu)
@@ -17,21 +17,21 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <metashell/iface/preprocessor_shell.hpp>
+#include <metashell/iface/macro_discovery.hpp>
 
-#include <string>
+#include <metashell/clang_binary.hpp>
 
 namespace metashell
 {
-  class preprocessor_shell_constant : public iface::preprocessor_shell
+  class macro_discovery_clang : public iface::macro_discovery
   {
   public:
-    explicit preprocessor_shell_constant(data::result result_);
+    explicit macro_discovery_clang(clang_binary clang_binary_);
 
-    virtual data::result precompile(const std::string&) override;
+    virtual std::string macros(const iface::environment& env_) override;
 
   private:
-    data::result _result;
+    clang_binary _clang_binary;
   };
 }