+++ lib/core/engine_constant.cpp
@@ -19,6 +19,7 @@
 #include <metashell/engine.hpp>
 #include <metashell/engine_constant.hpp>
 #include <metashell/header_discoverer_constant.hpp>
+#include <metashell/macro_discovery_constant.hpp>
 #include <metashell/preprocessor_shell_constant.hpp>
 #include <metashell/template_tracer_constant.hpp>
 #include <metashell/type_shell_constant.hpp>
@@ -54,7 +55,8 @@ std::unique_ptr<iface::engine> metashell::create_failing_engine()
       "failing", type_shell_constant(result),
       preprocessor_shell_constant(result), code_completer_constant(),
       header_discoverer_constant(empty, empty),
-      template_tracer_constant(result), cpp_validator_constant(result));
+      template_tracer_constant(result), cpp_validator_constant(result),
+      macro_discovery_constant());
 }
 
 std::unique_ptr<iface::engine>
@@ -67,7 +69,8 @@ metashell::create_engine_returning_type(const std::string& type_)
       "type_returning", type_shell_constant(result),
       preprocessor_shell_constant(result), code_completer_constant(),
       header_discoverer_constant(empty, empty),
-      template_tracer_constant(result), cpp_validator_constant(result));
+      template_tracer_constant(result), cpp_validator_constant(result),
+      macro_discovery_constant());
 }
 
 std::unique_ptr<iface::engine> metashell::create_engine_with_include_path(
@@ -79,5 +82,6 @@ std::unique_ptr<iface::engine> metashell::create_engine_with_include_path(
       "engine_with_include_path", type_shell_constant(result),
       preprocessor_shell_constant(result), code_completer_constant(),
       create_header_discoverer_with_include_path(type_, path_),
-      template_tracer_constant(result), cpp_validator_constant(result));
+      template_tracer_constant(result), cpp_validator_constant(result),
+      macro_discovery_constant());
 }