+++ lib/core/macro_discovery_clang.cpp
@@ -1,6 +1,3 @@
-#ifndef METASHELL_PREPROCESSOR_SHELL_CONSTANT_HPP
-#define METASHELL_PREPROCESSOR_SHELL_CONSTANT_HPP
-
 // Metashell - Interactive C++ template metaprogramming shell
 // Copyright (C) 2016, Abel Sinkovics (abel@sinkovics.hu)
 //
@@ -17,22 +14,28 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <metashell/iface/preprocessor_shell.hpp>
-
-#include <string>
+#include <metashell/macro_discovery_clang.hpp>
 
 namespace metashell
 {
-  class preprocessor_shell_constant : public iface::preprocessor_shell
+  macro_discovery_clang::macro_discovery_clang(clang_binary clang_binary_)
+    : _clang_binary(clang_binary_)
   {
-  public:
-    explicit preprocessor_shell_constant(data::result result_);
+  }
 
-    virtual data::result precompile(const std::string&) override;
+  std::string macro_discovery_clang::macros(const iface::environment& env_)
+  {
+    const data::process_output output =
+        run_clang(_clang_binary, {"-dM", "-E"}, env_.get_all());
 
-  private:
-    data::result _result;
-  };
+    if (output.exit_code == data::exit_code_t(0))
+    {
+      return output.standard_output;
+    }
+    else
+    {
+      throw std::runtime_error("Error getting list of macros: " +
+                               output.standard_error);
+    }
+  }
 }
-
-#endif