+++ lib/core/preprocessor_shell_vc.cpp
@@ -1,6 +1,3 @@
-#ifndef METASHELL_PREPROCESSOR_SHELL_CONSTANT_HPP
-#define METASHELL_PREPROCESSOR_SHELL_CONSTANT_HPP
-
 // Metashell - Interactive C++ template metaprogramming shell
 // Copyright (C) 2016, Abel Sinkovics (abel@sinkovics.hu)
 //
@@ -17,22 +14,22 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <metashell/iface/preprocessor_shell.hpp>
-
-#include <string>
+#include <metashell/preprocessor_shell_vc.hpp>
 
 namespace metashell
 {
-  class preprocessor_shell_constant : public iface::preprocessor_shell
+  preprocessor_shell_vc::preprocessor_shell_vc(vc_binary vc_binary_)
+    : _vc_binary(vc_binary_)
   {
-  public:
-    explicit preprocessor_shell_constant(data::result result_);
+  }
 
-    virtual data::result precompile(const std::string&) override;
+  data::result preprocessor_shell_vc::precompile(const std::string& exp_)
+  {
+    const data::process_output output = run_vc(_vc_binary, {"/E"}, exp_);
 
-  private:
-    data::result _result;
-  };
+    const bool success = output.exit_code == data::exit_code_t(0);
+
+    return data::result{success, success ? output.standard_output : "",
+                        success ? "" : vc_error_report_on_stderr(output), ""};
+  }
 }
-
-#endif