+++ lib/data/process_output.cpp
@@ -1,6 +1,3 @@
-#ifndef METASHELL_PREPROCESSOR_SHELL_CONSTANT_HPP
-#define METASHELL_PREPROCESSOR_SHELL_CONSTANT_HPP
-
 // Metashell - Interactive C++ template metaprogramming shell
 // Copyright (C) 2016, Abel Sinkovics (abel@sinkovics.hu)
 //
@@ -17,22 +14,27 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <metashell/iface/preprocessor_shell.hpp>
+#include <metashell/data/process_output.hpp>
 
-#include <string>
+#include <boost/algorithm/string/replace.hpp>
+
+namespace
+{
+  std::string dos2unix(std::string s_)
+  {
+    boost::algorithm::replace_all(s_, "\r\n", "\n");
+    return s_;
+  }
+}
 
 namespace metashell
 {
-  class preprocessor_shell_constant : public iface::preprocessor_shell
+  namespace data
   {
-  public:
-    explicit preprocessor_shell_constant(data::result result_);
-
-    virtual data::result precompile(const std::string&) override;
-
-  private:
-    data::result _result;
-  };
+    process_output dos2unix(process_output o_)
+    {
+      return {o_.exit_code, ::dos2unix(move(o_.standard_output)),
+              ::dos2unix(move(o_.standard_error))};
+    }
+  }
 }
-
-#endif